- fix handling of libmm deps - some spec cleanups
git-svn-id: svn+ssh://thor/srv/subversion/annvix/repos/packages@6896 f191e5f7-4e04-0410-925a-985e7dbff663

+++ removed/branches/CURRENT_1_0/unpackages/apache-suexec/apache-suexec.spec
@@ -1,12 +1,16 @@
 %define name	apache-suexec
 %define version %{apache_version}
-%define release 2sls
+%define release 3sls
 
 %{expand:%%define apache_version %(rpm -q apache-devel|sed 's/apache-devel-\([0-9].*\)-.*$/\1/')}
 %{expand:%%define apache_release %(rpm -q apache-devel|sed 's/apache-devel-[0-9].*-\(.*\)$/\1/')}
 
-%{expand:%%define mm_major %(mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\1/')}
-%{expand:%%define mm_minor %(mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\2/')}
+# the default version of mm required
+%define dmm_major	1
+%define dmm_minor	3.0
+
+%{expand:%%define mm_major %([ -x /usr/bin/mm-config ] && mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\1/' || echo "%{dmm_major}")}
+%{expand:%%define mm_minor %([ -x /usr/bin/mm-config ] && mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\2/' || echo "%{dmm_minor}")}
 %define mm_version %{mm_major}.%{mm_minor}
 
 Summary:	Suexec binary for apache
@@ -37,7 +41,7 @@ the same user who is running the web server.
 
 
 %prep
-%setup -n %{name}
+%setup -q -n %{name}
 
 APFLAGS=`/usr/sbin/apxs -q CFLAGS`
 OPTIONS=$APFLAGS" \
@@ -53,22 +57,26 @@ echo Configuring with: $OPTIONS
 gcc $OPTIONS -o suexec suexec.c
 
 %install
-rm -rf $RPM_BUILD_ROOT
+[ -n "%{buildroot}" -a "%{buildroot}" != / ] && rm -rf %{buildroot}
+
 mkdir -p %{buildroot}%{_sbindir}
 mkdir -p %{buildroot}%{_mandir}/man8
 install -m 4711 suexec %{buildroot}%{_sbindir}
 install suexec.8 %{buildroot}%{_mandir}/man8
 
+%clean
+[ -n "%{buildroot}" -a "%{buildroot}" != / ] && rm -rf %{buildroot}
 
 %files
 %defattr(-,root,root)
 %attr(4710,root,apache) %{_sbindir}/suexec
 %{_mandir}/man8/*
 
-%clean
-rm -rf %{buildroot}
-
 %changelog
+* Tue Feb 24 2004 Vincent Danen <vdanen@opensls.org> 1.3.29-3sls
+- fix handling of libmm deps
+- some spec cleanups
+
 * Sat Jan 04 2004 Vincent Danen <vdanen@opensls.org> 1.3.29-2sls
 - OpenSLS build
 - tidy spec
